+++ example/lib/examples/counter_app_with_error.dart
@@ -6,7 +6,7 @@ import 'package:states_rebuilder/states_rebuilder.dart';
 class Counter {
   int _count = 0;
   int get count => _count;
-  Future<void> increment() async {
+  void increment() async {
     //Simulating async task
     await Future<void>.delayed(const Duration(seconds: 1));
     //Simulating error (50% chance of error);
@@ -17,6 +17,10 @@ class Counter {
     }
     _count++;
   }
+
+  void dispose() {
+    print('dispose');
+  }
 }
 
 class App extends StatelessWidget {
@@ -24,8 +28,9 @@ class App extends StatelessWidget {
   Widget build(BuildContext context) {
     return Injector(
       inject: [Inject<Counter>(() => Counter())],
+      disposeModels: true,
       builder: (BuildContext context) {
-        final ReactiveModel<Counter> counterModel =
+        final ReactiveModel<Counter> counterModelRM =
             Injector.getAsReactive<Counter>();
         return Scaffold(
           appBar: AppBar(
@@ -34,17 +39,17 @@ class App extends StatelessWidget {
           body: MyHome(),
           floatingActionButton: FloatingActionButton(
             child: Icon(Icons.add),
-            onPressed: () => counterModel.setState(
+            onPressed: () => counterModelRM.setState(
               (Counter state) => state.increment(),
               catchError: true, //catch the error
               onSetState: (BuildContext context) {
                 // osSetState will be executed after mutating the state.
-                if (counterModel.hasError) {
+                if (counterModelRM.hasError) {
                   showDialog<dynamic>(
                     context: context,
                     builder: (BuildContext context) => AlertDialog(
                       title: const Text('Error!'),
-                      content: Text('${counterModel.error}'),
+                      content: Text('${counterModelRM.error}'),
                     ),
                   );
                 }