+++ lib/src/injector.dart
@@ -66,7 +66,11 @@ class Injector extends StatefulWidget {
   ///Called after the widget is inserted in the widget tree.
   final void Function(BuildContext context) afterInitialBuild;
 
-  ///Set to true to dispose all models.
+  ///if it is set to true all injected models will be disposed.
+  ///
+  ///Injected models are disposed by calling the 'dispose()' method if exists.
+  ///
+  ///In any of the injected classes you can define a 'dispose()' method to clean up resources.
   final bool disposeModels;
 
   /// get the same singleton
@@ -78,17 +82,12 @@ class Injector extends StatefulWidget {
     if (inject == null) {
       return null;
     }
-
-    assert(
-      () {
-        if (inject.isAsyncType == true) {
-          throw Exception(AssertMessage.getInjectStreamAndFutureError());
-        }
-        return true;
-      }(),
-    );
-
-    final T model = inject?.getSingleton();
+    T model;
+    if (inject.isAsyncType == true) {
+      model = inject?.getReactiveSingleton()?.state;
+    } else {
+      model = inject?.getSingleton();
+    }
 
     if (context != null) {
       assert(
@@ -129,23 +128,6 @@ class Injector extends StatefulWidget {
       }
     }
     return model;
-  }
-
-  ///Use [getAsReactive] instead. It will be removed in next releases.
-  @deprecated
-  static ReactiveModel<T> getAsModel<T>(
-      {dynamic name,
-      BuildContext context,
-      bool silent = false,
-      bool asNewReactiveInstance = false,
-      bool resetStateStatus = false}) {
-    return getAsReactive<T>(
-      name: name,
-      context: context,
-      silent: silent,
-      asNewReactiveInstance: asNewReactiveInstance,
-      keepCustomStateStatus: resetStateStatus,
-    );
   }
 
   ///Get The registered reactive singleton with type $T or with name [name] if it is defined.