+++ test/injector_test.dart
@@ -232,22 +232,51 @@ void main() {
   });
 
   testWidgets(
-      'Injector : should throw if get method is used with stream or future injection',
+      'Injector : should Injector.get work for model injected with Inject.Future',
       (WidgetTester tester) async {
     await tester.pumpWidget(
       Injector(
         inject: [
           Inject<bool>.future(
-              () => Future.delayed(Duration(seconds: 1), () => false)),
+            () => Future.delayed(
+              Duration(seconds: 1),
+              () => false,
+            ),
+            initialValue: true,
+            isLazy: false,
+          ),
         ],
         builder: (_) {
           return Container();
         },
       ),
     );
-    expect(() => Injector.get<bool>(), throwsException);
+    expect(Injector.get<bool>(), isTrue);
+    await tester.pump(Duration(seconds: 2));
+    expect(Injector.get<bool>(), isFalse);
   });
 
+  testWidgets('Injector : should Inject.future not lazily work',
+      (WidgetTester tester) async {
+    await tester.pumpWidget(
+      Injector(
+        inject: [
+          Inject<bool>.future(
+            () => Future.delayed(
+              Duration(seconds: 1),
+              () => false,
+            ),
+            isLazy: false,
+          ),
+        ],
+        builder: (_) {
+          return Container();
+        },
+      ),
+    );
+    await tester.pump(Duration(seconds: 2));
+    expect(Injector.getAsReactive<bool>().state, isFalse);
+  });
   testWidgets(
       'Injector : should throw if get method with context is used with non reactive model',
       (WidgetTester tester) async {
@@ -438,8 +467,8 @@ void main() {
         ),
       );
 
-      expect(() => Injector.getAsReactive<int>().setState((state) {}),
-          throwsException);
+      expect(
+          () => Injector.getAsReactive<int>().setState(null), throwsException);
     },
   );
 
@@ -458,8 +487,8 @@ void main() {
         ),
       );
 
-      expect(() => Injector.getAsReactive<int>().setState((state) {}),
-          throwsException);
+      expect(
+          () => Injector.getAsReactive<int>().setState(null), throwsException);
       await tester.pump(Duration(seconds: 1));
     },
   );
@@ -577,7 +606,7 @@ void main() {
 
     isRebuilt = false;
     //notify to rebuild
-    vm.setState((_) {});
+    vm.setState(null);
     await tester.pump();
     expect(Injector.get<ViewModel>().message, equals("I am injected"));
     expect(Injector.get<Service1>().message, equals("I am Service1"));
@@ -623,7 +652,7 @@ void main() {
     expect(disposeStateIsCalled, isFalse);
 
     switcher = false;
-    vm.setState((_) {});
+    vm.setState(null);
     await tester.pump();
     expect(Injector.get<ViewModel>().message, equals("I am injected"));
     expect(() => Injector.get<Service1>(), throwsException);
@@ -696,7 +725,7 @@ void main() {
     isRebuilt = false;
     expect(vm, isNot(isNull));
     //notify to rebuild
-    vm.setState((_) {});
+    vm.setState(null);
     await tester.pump();
     expect(Injector.get<ViewModel>().message, equals("I am injected"));
     expect(Injector.get<Service1>().message, equals("I am Service1"));
@@ -749,7 +778,7 @@ void main() {
       );
 
       expect(numberOfCall, 1);
-      vm.setState((_) {});
+      vm.setState(null);
       await tester.pump();
       expect(numberOfCall, 1);
     },
@@ -1094,7 +1123,7 @@ void main() {
     expect(numberOfRebuild, equals(1));
 
     switcher = false;
-    integerModel.setState((_) {});
+    integerModel.setState(null);
     await tester.pump();
     expect(numberOfRebuild, equals(2));
     expect(inject.newReactiveInstanceList.length, 0);
@@ -1125,7 +1154,7 @@ void main() {
     expect(numberOfRebuildSingleton, equals(1));
     expect(numberOfRebuildNewReactive, equals(1));
 
-    integerModelFromInsideStateBuilder.setState((_) {});
+    integerModelFromInsideStateBuilder.setState(null);
     await tester.pump();
     expect(numberOfRebuildSingleton, equals(1));
     expect(numberOfRebuildNewReactive, equals(2));
@@ -1153,7 +1182,7 @@ void main() {
     );
     expect(numberOfRebuild, equals(1));
 
-    integerModelFromInsideStateBuilder.setState((_) {});
+    integerModelFromInsideStateBuilder.setState(null);
     await tester.pump();
     expect(numberOfRebuild, equals(2));
   });
@@ -1183,7 +1212,7 @@ void main() {
     expect(reactiveSingleton.joinSingletonToNewData, isNull);
 
     newReactiveInstance.setState(
-      (_) {},
+      null,
       joinSingletonToNewData: 'From New reactive instance',
     );
     await tester.pump();
@@ -1191,7 +1220,7 @@ void main() {
         reactiveSingleton.joinSingletonToNewData, 'From New reactive instance');
 
     newReactiveInstance.setState(
-      (_) {},
+      null,
       joinSingletonToNewData: 'another message from new reactive instance',
     );
     await tester.pump();
@@ -1234,7 +1263,7 @@ void main() {
       expect(numberOfRebuild, equals(1));
       expect(numberOfRebuild1, equals(1));
       expect(numberOfRebuild2, equals(1));
-      integerModelFromInsideStateBuilder.setState((_) {},
+      integerModelFromInsideStateBuilder.setState(null,
           notifyAllReactiveInstances: true);
       await tester.pump();
       expect(numberOfRebuild, equals(1));
@@ -1299,7 +1328,7 @@ void main() {
       expect(integerModelFromInsideStateBuilder1.hasError, isTrue);
       expect(integerModelFromInsideStateBuilder2.hasError, isFalse);
 
-      integerModelFromInsideStateBuilder2.setState((_) {});
+      integerModelFromInsideStateBuilder2.setState(null);
       await tester.pump();
 
       expect(integerModelFromInsideStateBuilder.hasError, isTrue);
@@ -1364,7 +1393,7 @@ void main() {
       expect(integerModelFromInsideStateBuilder1.hasError, isTrue);
       expect(integerModelFromInsideStateBuilder2.hasError, isFalse);
 
-      integerModelFromInsideStateBuilder2.setState((_) {});
+      integerModelFromInsideStateBuilder2.setState(null);
       await tester.pump();
 
       expect(integerModelFromInsideStateBuilder.hasError, isFalse);
@@ -1429,7 +1458,7 @@ void main() {
       expect(integerModelFromInsideStateBuilder2.hasError, isFalse);
 
       integerModelFromInsideStateBuilder2.setState(
-        (_) {},
+        null,
         joinSingletonWith: JoinSingleton.withCombinedReactiveInstances,
       );
       await tester.pump();
@@ -1497,7 +1526,7 @@ void main() {
       expect(integerModelFromInsideStateBuilder2.hasError, isFalse);
 
       integerModelFromInsideStateBuilder2.setState(
-        (_) {},
+        null,
         joinSingletonWith: JoinSingleton.withNewReactiveInstance,
       );
       await tester.pump();
@@ -1533,7 +1562,7 @@ void main() {
         ),
       );
       model1.setState(
-        (_) {},
+        null,
         onSetState: (_) => _rebuildTracker.add('onSetState'),
         onRebuildState: (_) => _rebuildTracker.add('onRebuildState'),
       );
@@ -1659,10 +1688,10 @@ void main() {
       await tester.pump();
 
       expect(model1 == model2, isTrue);
-      model1.setState((_) {});
+      model1.setState(null);
       await tester.pump();
       expect(model1 == model2, isTrue);
-      model2.setState((_) {});
+      model2.setState(null);
       await tester.pump();
     },
   );
@@ -1701,10 +1730,10 @@ void main() {
       await tester.pump();
 
       expect(model1 == model2, isTrue);
-      model1.setState((_) {});
+      model1.setState(null);
       await tester.pump();
       expect(model1 == model2, isTrue);
-      model2.setState((_) {});
+      model2.setState(null);
       await tester.pump();
     },
   );
@@ -1758,20 +1787,77 @@ void main() {
         ),
       );
 
-      expect(model1.hashCode, equals(model1.hashCode));
+      expect(model1.hashCode, equals(model2.hashCode));
       expect(numberOFRebuild1, equals(1));
       expect(numberOFRebuild2, equals(1));
       isTrue = false;
       vm.rebuildStates();
       await tester.pump();
-      expect(model1.hashCode, equals(model1.hashCode));
+      expect(model1.hashCode, equals(model2.hashCode));
       expect(numberOFRebuild1, equals(1));
-      // expect(numberOFRebuild2, equals(2));//TODO build in initState of injector
-      model2.setState((_) {});
+      model2.setState(null);
       await tester.pump();
       expect(model1.hashCode, equals(model1.hashCode));
       expect(numberOFRebuild1, equals(1));
       expect(numberOFRebuild2, equals(2));
+    },
+  );
+
+  testWidgets(
+    'When a parent of injector rebuild the injector child tree will not rebuild',
+    (WidgetTester tester) async {
+      ReactiveModel<Integer> model1;
+      int numberOFRebuild1 = 0;
+
+      final vm = ViewModel();
+      await tester.pumpWidget(
+        StateBuilder(
+          models: [vm],
+          builder: (_, __) {
+            return Column(
+              children: <Widget>[
+                Injector(
+                  inject: [
+                    Inject<Integer>(() => Integer(0)),
+                  ],
+                  builder: (context) {
+                    model1 = Injector.getAsReactive<Integer>(context: context);
+                    numberOFRebuild1++;
+                    return Container();
+                  },
+                )
+              ],
+            );
+          },
+        ),
+      );
+
+      expect(numberOFRebuild1, equals(1));
+      vm.rebuildStates();
+      await tester.pump();
+      expect(numberOFRebuild1, equals(1));
+      model1.setState(null);
+      await tester.pump();
+      expect(numberOFRebuild1, equals(2));
+    },
+  );
+
+  testWidgets(
+    'should throw if setState is call on a reactive model with no subscribed widget',
+    (WidgetTester tester) async {
+      await tester.pumpWidget(
+        Injector(
+          inject: [
+            Inject<Integer>(() => Integer(0)),
+          ],
+          builder: (context) {
+            return Container();
+          },
+        ),
+      );
+
+      expect(() => Injector.getAsReactive<Integer>().setState(null),
+          throwsException);
     },
   );
 
@@ -1821,14 +1907,14 @@ void main() {
         ),
       );
 
-      model1.setState((state) {}, onSetState: (context) {
+      model1.setState(null, onSetState: (context) {
         context0 = context;
       });
       await tester.pump();
       expect(context0 == context2, isTrue);
       isTrue = false;
       vm.rebuildStates();
-      model1.setState((state) {}, onSetState: (context) {
+      model1.setState(null, onSetState: (context) {
         context0 = context;
       });
       await tester.pump();
@@ -1882,7 +1968,7 @@ void main() {
         ),
       );
 
-      model1.setState((state) {}, onRebuildState: (context) {
+      model1.setState(null, onRebuildState: (context) {
         context0 = context;
       });
       await tester.pump();
@@ -1890,7 +1976,7 @@ void main() {
 
       isTrue = false;
       vm.rebuildStates();
-      model1.setState((state) {}, onRebuildState: (context) {
+      model1.setState(null, onRebuildState: (context) {
         context0 = context;
       });
       await tester.pump();
@@ -1947,7 +2033,7 @@ void main() {
         ),
       );
 
-      model1.setState((state) {}, onRebuildState: (context) {
+      model1.setState(null, onRebuildState: (context) {
         context0 = context;
       });
       await tester.pump();
@@ -1955,7 +2041,7 @@ void main() {
 
       isTrue = false;
       vm.rebuildStates();
-      model1.setState((state) {}, onRebuildState: (context) {
+      model1.setState(null, onRebuildState: (context) {
         context0 = context;
       });
       await tester.pump();
@@ -2013,7 +2099,7 @@ void main() {
         ),
       );
 
-      model1.setState((state) {}, onRebuildState: (context) {
+      model1.setState(null, onRebuildState: (context) {
         context0 = context;
       });
       await tester.pump();
@@ -2021,7 +2107,7 @@ void main() {
 
       isTrue = false;
       vm.rebuildStates();
-      model1.setState((state) {}, onRebuildState: (context) {
+      model1.setState(null, onRebuildState: (context) {
         context0 = context;
       });
       await tester.pump();
@@ -2043,7 +2129,6 @@ void main() {
             ],
             builder: (context) {
               model = Injector.getAsReactive<Integer>(context: context);
-              print(model.snapshot);
               return model.whenConnectionState(
                 onIdle: () => Builder(
                   builder: (_) {
@@ -2076,15 +2161,78 @@ void main() {
       );
 
       expect(onConnectionState, equals('isIdle,'));
-      model.setState((state) => state.incrementAsync());
+      model.setState((state) {
+        return state.incrementAsync();
+      });
       await tester.pump();
       await tester.pump(Duration(seconds: 1));
       expect(onConnectionState, equals('isIdle, isWaiting, hasData'));
       onConnectionState = '';
-      model.setState((state) => state.incrementWithError(), catchError: true);
+      model.setState((state) => state.incrementWithError());
       await tester.pump();
       await tester.pump(Duration(seconds: 1));
       expect(onConnectionState, equals(' isWaiting, hasError'));
+    },
+  );
+
+  testWidgets(
+    "should string equality work : (== : true) (identical : false) (hashCode : true)",
+    (WidgetTester tester) async {
+      final s = {
+        'list': [1],
+      };
+      final s1 = s.toString();
+      final s2 = Map.from(s).toString();
+
+      final _equality = s1 == s2;
+      final _identical = identical(s1, s2);
+      final _hashCode = s1.hashCode == s2.hashCode;
+
+      expect(_equality, isTrue);
+      expect(_identical, isFalse);
+      expect(_hashCode, isTrue);
+    },
+  );
+
+  testWidgets(
+    "should ' StateBuilder with watch works",
+    (WidgetTester tester) async {
+      List<String> _rebuildTracker = [];
+      await tester.pumpWidget(
+        Injector(
+            inject: [Inject(() => Integer(0))],
+            builder: (context) {
+              return StateBuilder<Integer>(
+                models: [Injector.getAsReactive<Integer>()],
+                watch: (model) {
+                  return model.state.value;
+                },
+                builder: (_, __) {
+                  _rebuildTracker.add('rebuild');
+                  return Container();
+                },
+              );
+            }),
+      );
+      final model = Injector.getAsReactive<Integer>();
+      expect(_rebuildTracker, equals(['rebuild']));
+      model.setState(null);
+      await tester.pump();
+      expect(_rebuildTracker, equals(['rebuild']));
+
+      model.setState(null);
+      await tester.pump();
+      expect(_rebuildTracker, equals(['rebuild']));
+
+      model.state.value++;
+
+      model.setState(null);
+      await tester.pump();
+      expect(_rebuildTracker, equals(['rebuild', 'rebuild']));
+
+      model.setState(null);
+      await tester.pump();
+      expect(_rebuildTracker, equals(['rebuild', 'rebuild']));
     },
   );
 }
@@ -2120,12 +2268,12 @@ abstract class IService2 {
 class Integer {
   int value = 0;
   Integer(this.value);
-  incrementAsync() async {
-    await Future.delayed(Duration(seconds: 1));
+  void incrementAsync() async {
+    await Future.delayed(Duration(microseconds: 1));
     value++;
   }
 
-  incrementWithError() async {
+  void incrementWithError() async {
     await Future.delayed(Duration(seconds: 1));
     throw Exception('There is an error');
   }