+++ CHANGELOG.md
@@ -1,5 +1,12 @@
+## 1.11.0 (2020-01-05)
+* `Inject.get` for injected streams and future will no longer throw, it will return the the current value.
+* If `whenConnectionState` is defined, `catchError` is set true automatically.
+* Add `watch` parameter to `StateBuilder` widget and. `watch` allows to link the rebuild process to the variation of a set of variables.(Experimental feature).
+* Remove deprecated `getAsModel` and `hasState`.
+* Update docs and add Dependency Injection section in the readme file.
+
 ## 1.10.0 (2019-12-30)
-* Add `whenConnectionState` method to the `ReactiveModel`. IT exhaustively switch over all the possible statuses of `connectionState`. Used mostly to return a Widget.
+* Add `whenConnectionState` method to the `ReactiveModel`. IT exhaustively switch over all the possible statuses of `connectionState`. Used mostly to return a Widget. (Pul request of [ResoCoder](https://resocoder.com/2019/12/30/states-rebuilder-zero-boilerplate-flutter-state-management/)).
 
 ## 1.9.0 (2019-12-28)
 * Add assertion error helpful messages.